requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argilla==2.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
datasets==3.1.0
dill==0.3.8
distilabel==1.4.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
propcache==0.2.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scipy==1.14.1
setuptools==75.5.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tblib==3.0.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
universal_pathlib==0.2.5
urllib3==2.2.3
xxhash==3.5.0
yarl==1.17.1
